Move organization and user forms to shared and add components to settings module
