+++ src/app/components/index.ts
@@ -1,20 +1,16 @@
 export * from './about-us/about-us.component';
 export * from './check-in-form/check-in-form.component';
 export * from './signature-field/signature-field.component';
-export * from '../shared/components/daily-hours-chart/daily-hours-chart.component';
-export * from '../shared/components/data-cell/data-cell.component';
-export * from '../shared/components/data-table/data-table.component';
 export * from './header/header.component';
 export * from './new-volunteer-form/new-volunteer-form.component';
 export * from './organization-confirm/organization-confirm.component';
-export * from './organization-form/organization-form.component';
-export * from './reports-form/reports-form.component';
+export * from '../shared/components/organization-form/organization-form.component';
+export * from '../settings/components/reports-form/reports-form.component';
 export * from './sidenav/sidenav.component';
-export * from './user-form/user-form.component';
+export * from '../shared/components/user-form/user-form.component';
 export * from './volunteer-menu/volunteer-menu.component';
 export * from './password-dialog/password-dialog.component';
 export * from './email-dialog/email-dialog.component';
-export * from '../shared/components/data-display/data-display.component';
 export * from './services-agreement/services-agreement.component';
 export * from './services-agreement-dialog/services-agreement-dialog.component';
 export * from './loader/loader.component';
