+++ src/app/settings/components/reports-form/reports-form.component.ts
@@ -3,7 +3,7 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MatSlideToggle } from '@angular/material';
 import { Subscription } from 'rxjs';
 
-import { Report } from '../../models';
+import { Report } from '../../../models/index';
 
 @Component({
   selector: 'app-reports-form',
