+++ src/app/shared/components/user-form/user-form.component.spec.ts
@@ -3,7 +3,7 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { MatIconModule, MatInputModule } from '@angular/material';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 
-import { getTestUsers, User } from '../../models';
+import { getTestUsers, User } from '../../../models/index';
 import { UserFormComponent } from './user-form.component';
 
 describe('UserFormComponent', () => {
