+++ src/app/shared/shared.module.ts
@@ -1,10 +1,25 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
+import { MatIconModule, MatInputModule } from '@angular/material';
+import { OrganizationFormComponent } from './components/organization-form/organization-form.component';
+import { UserFormComponent } from './components/user-form/user-form.component';
 
 @NgModule({
   imports: [
-    CommonModule
+    CommonModule,
+    MatIconModule,
+    MatInputModule,
+    ReactiveFormsModule,
   ],
-  declarations: []
+  declarations: [
+    OrganizationFormComponent,
+    UserFormComponent,
+  ],
+  exports: [
+    OrganizationFormComponent,
+    UserFormComponent,
+  ],
 })
-export class SharedModule { }
+export class SharedModule {
+}
